api/pyproject.toml
[project]
name = "strot-api"
version = "0.0.1"
description = "API server for strot"
authors = [{ name = "Harsh Verma", email = "harsh@vertexcover.io" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "alembic>=1.16.3",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.0",
    "json-schema-to-pydantic>=0.4.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["strot_api*"]
